+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api/ChatApi.kt
@@ -23,7 +23,6 @@ import io.getstream.chat.android.client.api.models.QueryChannelRequest
 import io.getstream.chat.android.client.api.models.QueryChannelsRequest
 import io.getstream.chat.android.client.api.models.QueryThreadsRequest
 import io.getstream.chat.android.client.api.models.QueryUsersRequest
-import io.getstream.chat.android.client.api.models.SearchMessagesRequest
 import io.getstream.chat.android.client.api.models.SendActionRequest
 import io.getstream.chat.android.client.events.ChatEvent
 import io.getstream.chat.android.client.utils.ProgressCallback
@@ -179,23 +178,7 @@ internal interface ChatApi {
     ): Call<List<Reaction>>
 
     @CheckResult
-    fun sendReaction(reaction: Reaction, enforceUnique: Boolean): Call<Reaction>
-
-    @CheckResult
-    fun sendReaction(
-        messageId: String,
-        reactionType: String,
-        enforceUnique: Boolean,
-    ): Call<Reaction> {
-        return sendReaction(
-            reaction = Reaction(
-                messageId = messageId,
-                type = reactionType,
-                score = 0,
-            ),
-            enforceUnique = enforceUnique,
-        )
-    }
+    fun sendReaction(reaction: Reaction, enforceUnique: Boolean, skipPush: Boolean): Call<Reaction>
 
     @CheckResult
     fun deleteReaction(messageId: String, reactionType: String): Call<Message>