+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/api2/mapping/DtoMappingTest.kt
@@ -228,6 +228,7 @@ internal class DtoMappingTest {
             user = reaction.user?.let { with(mapping) { it.toDto() } },
             user_id = reaction.userId,
             extraData = reaction.extraData,
+            emoji_code = reaction.emojiCode,
         )
         dto shouldBeEqualTo expected
     }