+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Add support for `message.updated` and `reaction.new` push notifications. [#5935](https://github.com/GetStream/stream-chat-android/pull/5935)
+- Add `skipPush` parameter to `ChatClient.sendReaction` to skip sending push notifications for the reaction. [#5935](https://github.com/GetStream/stream-chat-android/pull/5935)
 - Add `ChatClient.setUserPushPreference(level: PushPreferenceLevel)` for setting the global (user) preferences. [#5942](https://github.com/GetStream/stream-chat-android/pull/5942)
 - Add `ChatClient.setChannelPushPreference(cid: String, level: PushPreferenceLevel)` for setting the channel-specific preferences.  [#5942](https://github.com/GetStream/stream-chat-android/pull/5942)
 - Add `ChatClient.snoozeUserPushNotifications(until: Date)` for global (user) snoozing of push notifications. [#5942](https://github.com/GetStream/stream-chat-android/pull/5942)
@@ -55,6 +57,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Add `ReactionPushEmojiFactory` for generating emoji codes for reaction push notifications. [#5935](https://github.com/GetStream/stream-chat-android/pull/5935)
 
 ### ⚠️ Changed
 
@@ -66,6 +69,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Add `ChatUi.reactionPushEmojiFactory` for customizing the emoji codes for reaction push notifications. [#5935](https://github.com/GetStream/stream-chat-android/pull/5935)
 
 ### ⚠️ Changed
 
@@ -77,6 +81,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Add `ChatTheme.reactionPushEmojiFactory` for customizing the emoji codes for reaction push notifications. [#5935](https://github.com/GetStream/stream-chat-android/pull/5935)
 
 ### ⚠️ Changed